LLM_training/requirements.txt
# Core ML libraries
torch>=2.0.0
transformers>=4.35.0
peft>=0.6.0
datasets>=2.14.0
accelerate>=0.24.0

# Training utilities
wandb>=0.16.0
tensorboard>=2.14.0
tqdm>=4.66.0

# Data processing
pandas>=2.1.0
numpy>=1.24.0
scikit-learn>=1.3.0

# Configuration and utilities
pyyaml>=6.0
omegaconf>=2.3.0
hydra-core>=1.3.0

# Evaluation metrics
evaluate>=0.4.0
rouge-score>=0.1.2
sacrebleu>=2.3.1

# Memory optimization
bitsandbytes>=0.41.0

# Development tools
jupyter>=1.0.0
ipywidgets>=8.1.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Optional: For specific model types
sentencepiece>=0.1.99
protobuf>=4.24.0

# For chat/instruction datasets
jinja2>=3.1.0 